+++ demo/app/login/login.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { MsalService } from 'modules';
 import { UserService } from './../shared/services/user.service';
 import { Router } from '@angular/router';
@@ -9,26 +9,27 @@ import { StringDict } from 'msal/lib-commonjs/MsalTypes';
   selector: 'app-login',
   templateUrl: './login.component.html'
 })
-export class LoginComponent {
+export class LoginComponent implements OnInit {
 
 
   constructor(private msalService: MsalService, public userService: UserService, private router: Router) { }
 
+  ngOnInit(): void {
+    this.msalService.logout();
+  }
+
   public async login(popup = false) {
-    this.msalService.authority = environment.authority + environment.userflow;
-    const key = 'mkt';
-    const qp: StringDict = {};
-    qp[key] = 'en-US';
     if (!popup) {
-      this.msalService.loginRedirect({
-        authority: environment.authority + environment.userflow,
-        extraQueryParameters: qp
-      });
+      // EXTRA PARAMETERS CAN BE PASSED TO AUTHENTICATE AGAINS ANOTHER AUTHORITY 
+      // OR PASS A DIFFERENT LANGUAGE.
+      // ------------------ example ----------------------
+      // this.msalService.authority = environment.authority + environment.userflow;
+      // const key = 'mkt';
+      // const qp: StringDict = {};
+      // qp[key] = 'en-US';  
+      this.msalService.loginRedirect();
     } else {
-      const response = await this.msalService.loginPopup({
-        authority: environment.authority + environment.userflow,
-        extraQueryParameters: qp
-      });
+      const response = await this.msalService.loginPopup();
       if (response.idToken) {
         this.router.navigate(['/my-profile']);
       }