+++ modules/msal-config.ts
@@ -1,24 +1,23 @@
 import { LogLevel, CacheLocation } from 'msal';
-import { ILoggerCallback } from 'msal/lib-commonjs/Logger';
-import { StringDict } from 'msal/lib-commonjs/MsalTypes';
 
 export class MsalConfig {
-    clientID: string;
-    authority ?= 'https://login.microsoftonline.com/common';
+    userinteractionInGuard?= false;
+    lang?: string;
+
+    clientId: string;
+    authority?: string;
     tokenReceivedCallback?: boolean;
-    validateAuthority ?= true;
-    storeAuthStateInCookie ?= true;
+    validateAuthority?= true;
+    storeAuthStateInCookie?= true;
     cacheLocation?: CacheLocation;
     redirectUri?: string | (() => string);
     postLogoutRedirectUri?: string | (() => string);
-    logger?: ILoggerCallback;
-    loadFrameTimeout ?= 6000;
-    navigateToLoginRequestUrl ?= false;
+    loadFrameTimeout?: number;
+    navigateToLoginRequestUrl?= false;
     popUp?: boolean;
     consentScopes?: string[];
     unprotectedResources?: string[] = [];
     protectedResourceMap?: [string, string[]][] = [];
-    extraQueryParameters?: StringDict;
     correlationId?: string;
     level?: LogLevel;
     piiLoggingEnabled?: boolean;